+++ Cargo.toml
@@ -3,6 +3,7 @@ resolver = "2"
 members = [
     "xrpl-wasm-stdlib",
     "xrpl-wasm-stdlib/xrpl-macros",
+    "xrpl-escrow-stdlib",
 ]
 exclude = [
     "examples",
