+++ examples/smart-escrows/atomic_swap/atomic_swap1/src/lib.rs
@@ -4,9 +4,6 @@
 extern crate std;
 
 use xrpl_wasm_stdlib::core::keylets::XRPL_KEYLET_SIZE;
-use xrpl_wasm_stdlib::core::ledger_objects::current_escrow::{self, CurrentEscrow};
-use xrpl_wasm_stdlib::core::ledger_objects::escrow::Escrow;
-use xrpl_wasm_stdlib::core::ledger_objects::traits::{CurrentEscrowFields, EscrowFields};
 use xrpl_wasm_stdlib::core::locator::Locator;
 use xrpl_wasm_stdlib::core::types::contract_data::XRPL_CONTRACT_DATA_SIZE;
 use xrpl_wasm_stdlib::host;
@@ -17,6 +14,10 @@ use xrpl_wasm_stdlib::host::trace::{DataRepr, trace_data, trace_num};
 use xrpl_wasm_stdlib::host::{Error, Result, Result::Err, Result::Ok};
 use xrpl_wasm_stdlib::sfield;
 use xrpl_wasm_stdlib::types::{ContractData, XRPL_CONTRACT_DATA_SIZE as TX_CONTRACT_DATA_SIZE};
+
+use xrpl_escrow_stdlib::ledger_objects::current_escrow::{self, CurrentEscrow};
+use xrpl_escrow_stdlib::ledger_objects::escrow::Escrow;
+use xrpl_escrow_stdlib::ledger_objects::traits::{CurrentEscrowFields, EscrowFields};
 
 // Security constants for validation
 const VALIDATION_FAILED: i32 = 0;
