+++ examples/smart-escrows/atomic_swap/atomic_swap2/Cargo.toml
@@ -12,3 +12,4 @@ crate-type = ["cdylib"]
 
 [dependencies]
 xrpl-wasm-stdlib = { path = "../../../../xrpl-wasm-stdlib" }
+xrpl-escrow-stdlib = { path = "../../../../xrpl-escrow-stdlib" }
