+++ examples/smart-escrows/freelancer_escrow/Cargo.toml
@@ -8,3 +8,4 @@ crate-type = ["cdylib"]
 
 [dependencies]
 xrpl-wasm-stdlib = { path = "../../../xrpl-wasm-stdlib" }
+xrpl-escrow-stdlib = { path = "../../../xrpl-escrow-stdlib" }
