+++ examples/smart-escrows/freelancer_escrow/src/lib.rs
@@ -3,10 +3,7 @@
 #[cfg(not(target_arch = "wasm32"))]
 extern crate std;
 
-use xrpl_wasm_stdlib::core::current_tx::escrow_finish::get_current_escrow_finish;
 use xrpl_wasm_stdlib::core::current_tx::traits::TransactionCommonFields;
-use xrpl_wasm_stdlib::core::ledger_objects::current_escrow::{CurrentEscrow, get_current_escrow};
-use xrpl_wasm_stdlib::core::ledger_objects::traits::CurrentEscrowFields;
 use xrpl_wasm_stdlib::core::locator::Locator;
 use xrpl_wasm_stdlib::core::types::account_id::AccountID;
 use xrpl_wasm_stdlib::core::types::contract_data::ContractData;
@@ -15,6 +12,10 @@ use xrpl_wasm_stdlib::host::get_tx_nested_field;
 use xrpl_wasm_stdlib::host::trace::trace_num;
 use xrpl_wasm_stdlib::host::{Error, Result, Result::Err, Result::Ok};
 use xrpl_wasm_stdlib::sfield;
+
+use xrpl_escrow_stdlib::current_tx::escrow_finish::get_current_escrow_finish;
+use xrpl_escrow_stdlib::ledger_objects::current_escrow::{CurrentEscrow, get_current_escrow};
+use xrpl_escrow_stdlib::ledger_objects::traits::CurrentEscrowFields;
 
 macro_rules! try_or_trace {
     ($e:expr, $label:literal) => {
