+++ examples/smart-escrows/kyc/Cargo.toml
@@ -10,3 +10,4 @@ crate-type = ["cdylib"]
 
 [dependencies]
 xrpl-wasm-stdlib = { path = "../../../xrpl-wasm-stdlib" }
+xrpl-escrow-stdlib = { path = "../../../xrpl-escrow-stdlib" }
