+++ xrpl-escrow-stdlib/src/current_tx/escrow_finish.rs
@@ -1,0 +1,16 @@
+use xrpl_wasm_stdlib::core::current_tx::traits::TransactionCommonFields;
+
+use crate::current_tx::traits::EscrowFinishFields;
+
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+#[repr(C)]
+pub struct EscrowFinish;
+
+impl TransactionCommonFields for EscrowFinish {}
+
+impl EscrowFinishFields for EscrowFinish {}
+
+#[inline]
+pub fn get_current_escrow_finish() -> EscrowFinish {
+    EscrowFinish
+}
