+++ xrpl-escrow-stdlib/src/ledger_objects/current_escrow.rs
@@ -1,4 +1,6 @@
-use crate::core::ledger_objects::traits::{CurrentEscrowFields, CurrentLedgerObjectCommonFields};
+use xrpl_wasm_stdlib::core::ledger_objects::traits::CurrentLedgerObjectCommonFields;
+
+use crate::ledger_objects::traits::CurrentEscrowFields;
 
 #[derive(Debug, Clone, Copy, Eq, PartialEq)]
 #[repr(C)]
