+++ xrpl-wasm-stdlib/src/core/mod.rs
@@ -8,8 +8,8 @@
 //! - [`types`]: Strongly-typed XRPL primitives (AccountID, Hash256, Amount, etc.)
 //! - [`locator`]: Build locators for nested field access
 //!
-//! Start with [`current_tx::escrow_finish::EscrowFinish`] to access EscrowFinish TX fields,
-//! or [`ledger_objects::current_escrow::get_current_escrow`] to access the active escrow.
+//! Escrow-specific wrappers (`EscrowFinish`, `CurrentEscrow`, `Escrow`) live in the
+//! companion `xrpl-escrow-stdlib` crate.
 
 pub mod constants;
 pub mod current_tx;
